projects/fmo_project/requirements.txt
amazon-braket-sdk==1.94.0
matplotlib==3.10.3
numpy==2.2.6
pandas==2.3.0
scipy==1.15.3
